requirements.txt
flask==1.0.2
torch==1.0.0
torchvision==0.2.1
numpy==1.15.4
Pillow==5.3.0
gunicorn==19.9.0